@@ -1,7 +1,5 @@
-'use strict';
-
-var m = require('mithril');
-var Utils = require('./utils');
+import m from 'mithril';
+import Utils from './utils';
 
 function getZero () {
   return '0';
@@ -136,4 +134,4 @@ function Operand () {
   return operand;
 }
 
-module.exports = Operand;
+export default Operand;
